tests/test_indicators.py
"""
Unit test untuk modul indicators: EMA dan CCI.
Menggunakan data sintetis untuk validasi kebenaran kalkulasi.
"""
from __future__ import annotations

import pandas as pd
import numpy as np
import pytest

from src.indicators.ema import calculate_ema, add_ema_columns
from src.indicators.cci import calculate_cci, add_cci_column


# ──────────────────────────────────────────────
# Fixtures
# ──────────────────────────────────────────────

@pytest.fixture
def sample_ohlcv_df() -> pd.DataFrame:
    """Buat DataFrame OHLCV sintetis dengan 50 candle."""
    np.random.seed(42)
    n = 50
    close_prices = 100 + np.cumsum(np.random.randn(n) * 0.5)
    high_prices = close_prices + np.abs(np.random.randn(n) * 0.3)
    low_prices = close_prices - np.abs(np.random.randn(n) * 0.3)
    open_prices = close_prices + np.random.randn(n) * 0.2
    volumes = np.random.randint(100, 10000, n).astype(float)

    timestamps = pd.date_range("2026-01-01", periods=n, freq="15min")

    df = pd.DataFrame({
        "open": open_prices,
        "high": high_prices,
        "low": low_prices,
        "close": close_prices,
        "volume": volumes,
    }, index=timestamps)

    return df


@pytest.fixture
def small_df() -> pd.DataFrame:
    """DataFrame kecil (5 candle) untuk test edge cases."""
    timestamps = pd.date_range("2026-01-01", periods=5, freq="15min")
    df = pd.DataFrame({
        "open": [100, 101, 102, 101, 103],
        "high": [101, 102, 103, 102, 104],
        "low": [99, 100, 101, 100, 102],
        "close": [100.5, 101.5, 102.5, 101.0, 103.5],
        "volume": [1000, 1100, 1200, 900, 1300],
    }, index=timestamps)
    return df


# ──────────────────────────────────────────────
# Test EMA
# ──────────────────────────────────────────────

class TestEMA:
    """Test suite untuk modul EMA."""

    def test_calculate_ema_basic(self, sample_ohlcv_df: pd.DataFrame) -> None:
        """EMA harus mengembalikan Series dengan panjang yang sama."""
        result = calculate_ema(sample_ohlcv_df, length=10)
        assert isinstance(result, pd.Series)
        assert len(result) == len(sample_ohlcv_df)
        assert result.name == "ema_10"

    def test_calculate_ema_not_nan_after_warmup(self, sample_ohlcv_df: pd.DataFrame) -> None:
        """EMA tidak boleh NaN setelah cukup data warmup."""
        result = calculate_ema(sample_ohlcv_df, length=10)
        # Setelah 10 candle pertama, tidak boleh ada NaN
        non_nan = result.dropna()
        assert len(non_nan) > 0
        assert len(non_nan) >= len(sample_ohlcv_df) - 10

    def test_calculate_ema_insufficient_data(self, small_df: pd.DataFrame) -> None:
        """EMA dengan data kurang dari length → semua NaN."""
        result = calculate_ema(small_df, length=20)
        assert result.isna().all()

    def test_calculate_ema_invalid_source(self, sample_ohlcv_df: pd.DataFrame) -> None:
        """EMA dengan source yang tidak ada → ValueError."""
        with pytest.raises(ValueError, match="tidak ditemukan"):
            calculate_ema(sample_ohlcv_df, length=10, source="invalid")

    def test_add_ema_columns(self, sample_ohlcv_df: pd.DataFrame) -> None:
        """add_ema_columns harus menambahkan 2 kolom EMA."""
        df = add_ema_columns(sample_ohlcv_df.copy(), fast=10, slow=20)
        assert "ema_10" in df.columns
        assert "ema_20" in df.columns

    def test_ema_fast_reacts_faster(self, sample_ohlcv_df: pd.DataFrame) -> None:
        """EMA fast (10) harus lebih reaktif dari EMA slow (20)."""
        df = add_ema_columns(sample_ohlcv_df.copy(), fast=10, slow=20)
        # EMA fast seharusnya lebih close ke harga terkini
        last_close = df["close"].iloc[-1]
        ema_10_diff = abs(df["ema_10"].iloc[-1] - last_close)
        ema_20_diff = abs(df["ema_20"].iloc[-1] - last_close)
        # Tidak selalu true tapi umumnya EMA fast lebih dekat
        # Kita hanya cek keduanya punya nilai valid
        assert not pd.isna(df["ema_10"].iloc[-1])
        assert not pd.isna(df["ema_20"].iloc[-1])


# ──────────────────────────────────────────────
# Test CCI
# ──────────────────────────────────────────────

class TestCCI:
    """Test suite untuk modul CCI."""

    def test_calculate_cci_basic(self, sample_ohlcv_df: pd.DataFrame) -> None:
        """CCI harus mengembalikan Series dengan panjang yang sama."""
        result = calculate_cci(sample_ohlcv_df, length=20)
        assert isinstance(result, pd.Series)
        assert len(result) == len(sample_ohlcv_df)
        assert result.name == "cci_20"

    def test_calculate_cci_not_nan_after_warmup(self, sample_ohlcv_df: pd.DataFrame) -> None:
        """CCI tidak boleh NaN setelah cukup data warmup."""
        result = calculate_cci(sample_ohlcv_df, length=20)
        non_nan = result.dropna()
        assert len(non_nan) > 0

    def test_calculate_cci_insufficient_data(self, small_df: pd.DataFrame) -> None:
        """CCI dengan data kurang dari length → semua NaN."""
        result = calculate_cci(small_df, length=20)
        assert result.isna().all()

    def test_calculate_cci_missing_columns(self) -> None:
        """CCI tanpa kolom high/low/close → ValueError."""
        df = pd.DataFrame({"close": [1, 2, 3]})
        with pytest.raises(ValueError, match="tidak ditemukan"):
            calculate_cci(df, length=20)

    def test_add_cci_column(self, sample_ohlcv_df: pd.DataFrame) -> None:
        """add_cci_column harus menambahkan kolom cci_20."""
        df = add_cci_column(sample_ohlcv_df.copy(), length=20)
        assert "cci_20" in df.columns

    def test_cci_oscillator_range(self, sample_ohlcv_df: pd.DataFrame) -> None:
        """CCI umumnya berada dalam range -200 hingga +200 (tapi bisa lebih)."""
        result = calculate_cci(sample_ohlcv_df, length=20)
        non_nan = result.dropna()
        # CCI bisa di luar -200/+200, tapi mayoritas harus wajar.
        # Pada data sintetis dengan low variance, pandas-ta bisa menghasilkan nilai ekstrem.
        assert non_nan.abs().mean() < 10000  # Sanity check yang lebih longgar
